security added in emptyInventory

+++ tunneler.lua
@@ -321,10 +321,12 @@ function emptyInventory() --if needed, empties the robots inventory in the chest
         for i = 1,16 do
             robot.select(i)
             local item = component.inventory_controller.getStackInInternalSlot(i)
-            if item.name ~= "minecraft:cobblestone" then
-                robot.drop()
-            else
-                robot.dropDown()
+            if item then
+                if item.name ~= "minecraft:cobblestone" then
+                    robot.drop()
+                else
+                    robot.dropDown()
+                end
             end
         end
         robot.turnLeft()
